@@ -8,19 +8,19 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "chinleung/laravel-locales": "^2.0",
         "chinleung/laravel-multilingual-routes": "^4.0",
         "doctrine/dbal": "^3.7",
         "guzzlehttp/guzzle": "^7.0.1",
-        "intervention/image": "^2.5",
-        "intervention/imagecache": "^2.4",
-        "laravel/framework": "^10.0",
-        "laravel/sanctum": "^3.2",
+        "intervention/image": "^3.0",
+        "intervention/image-laravel": "^1.0",
+        "laravel/framework": "^11.0",
+        "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2.0",
         "laravel/ui": "^4.0",
         "nesbot/carbon": "^2.36",
-        "opcodesio/log-viewer": "^2.0",
+        "opcodesio/log-viewer": "^3.0",
         "openai-php/client": "^0.2.0",
         "spatie/laravel-model-flags": "^1.0",
         "spatie/laravel-sluggable": "^3.4",
@@ -29,7 +29,7 @@
     "require-dev": {
         "spatie/laravel-ignition": "^2.0",
         "mockery/mockery": "^1.5",
-        "nunomaduro/collision": "^7.0",
+        "nunomaduro/collision": "^8.0",
         "phpunit/phpunit": "^10.0"
     },
     "config": {
