@@ -3,23 +3,23 @@ namespace App\Filters\Image\Template;
 use Intervention\Image\Interfaces\ImageInterface;
 use Intervention\Image\Interfaces\ModifierInterface;
 
-class Large implements ModifierInterface
+class Lightbox implements ModifierInterface
 {
     /**
      * Maximum width for large landscape images
      */    
-    protected $max_width = 1600;    
+    protected $max_width = 2000;    
 
     /**
      * Maximum height for large portrait images
      */    
-    protected $max_height = 900;
-    
+    protected $max_height = 1500;
+
     protected $maxSize;
 
     public function __construct($maxSize = NULL, $coords = FALSE, $ratio = NULL)
     {
-        $this->maxSize = $maxSize ? $maxSize : 1500;
+        $this->maxSize = $maxSize ? $maxSize : 2000;
     }
 
     public function apply(ImageInterface $image): ImageInterface
