@@ -1,43 +1,44 @@
 <?php
 namespace App\Filters\Image\Template;
-use Intervention\Image\Image;
-use Intervention\Image\Filters\FilterInterface;
+use Intervention\Image\Interfaces\ImageInterface;
+use Intervention\Image\Interfaces\ModifierInterface;
 
-class Small implements FilterInterface
+class Small implements ModifierInterface
 {
-  
-  /**
-   * Maximum width for small landscape images
-   */    
-  
-  protected $max_width = 800;    
+    /**
+     * Maximum width for small landscape images
+     */    
+    protected $max_width = 800;    
 
-  /**
-   * Maximum height for small portrait images
-   */
+    /**
+     * Maximum height for small portrait images
+     */
+    protected $max_height = 450;
+    
+    protected $maxSize;
 
-  protected $max_height = 450;
-  
-  public function applyFilter(Image $image)
-  {
-    // Get width and height
-    $width  = $image->getWidth();
-    $height = $image->getHeight();
-
-    // Resize landscape image
-    if ($width > $height && $width >= $this->max_width)
+    public function __construct($maxSize = NULL, $coords = FALSE, $ratio = NULL)
     {
-      $image->resize($this->max_width, null, function ($constraint) {
-        return $constraint->aspectRatio();
-      });
-    }
-    else if ($height >= $this->max_height)
-    {
-      $image->resize(null, $this->max_height, function ($constraint) {
-        return $constraint->aspectRatio();
-      });
+        $this->maxSize = $maxSize ? $maxSize : 800;
     }
 
-    return $image;
-  }
+    public function apply(ImageInterface $image): ImageInterface
+    {
+        // Get width and height
+        $width  = $image->width();
+        $height = $image->height();
+
+        // For landscape images
+        if ($width > $height && $width >= $this->max_width)
+        {
+            return $image->scaleDown(width: $this->max_width);
+        }
+        // For portrait images
+        else if ($height >= $this->max_height)
+        {
+            return $image->scaleDown(height: $this->max_height);
+        }
+
+        return $image;
+    }
 }
