@@ -1,66 +1,109 @@
 <?php
 namespace App\Http\Controllers;
-use Intervention\Image\ImageCacheController;
+use App\Facades\ImageCache;
 use Intervention\Image\ImageManager;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Response;
 use Config;
 
-class ImageController extends ImageCacheController
+class ImageController extends Controller
 {
-  protected $maxSize;
-  protected $coords;
-  protected $ratio;
-  
-  /**
-   * Get HTTP response of either original image file or
-   * template applied file.
-   *
-   * @param  string $template
-   * @param  string $filename
-   * @return Illuminate\Http\Response
-   */
+    protected $maxSize;
+    protected $coords;
+    protected $ratio;
+    
+    /**
+     * Get HTTP response of either original image file or
+     * template applied file.
+     *
+     * @param  string $template
+     * @param  string $filename
+     * @param  string|null $maxSize
+     * @param  string|null $coords
+     * @param  bool $ratio
+     * @return \Illuminate\Http\Response
+     */
+    public function getResponse($template, $filename, $maxSize = NULL, $coords = NULL, $ratio = FALSE)
+    {
+        $this->maxSize = $maxSize;
+        $this->coords = $coords;
+        $this->ratio = $ratio;
 
-  public function getResponse($template, $filename, $maxSize = NULL, $coords = NULL, $ratio = FALSE)
-  {
-    $this->maxSize  = $maxSize;
-    $this->coords   = $coords;
-    $this->ratio    = $ratio;
+        switch (strtolower($template)) {
+            case 'original':
+                return $this->getOriginal($filename);
 
-    switch (strtolower($template)) {
-      case 'original':
-        return $this->getOriginal($filename);
+            case 'download':
+                return $this->getDownload($filename);
 
-      case 'download':
-        return $this->getDownload($filename);
-
-      default:
-        return $this->getImage($template, $filename);
+            default:
+                return $this->getImage($template, $filename);
+        }
     }
-  }
 
-  /**
-   * Returns corresponding template object from given template name
-   *
-   * @param  string $template
-   * @return mixed
-   */
-  protected function getTemplate($template)
-  {
-    $template = config("imagecache.templates.{$template}");
+    /**
+     * Get cached image by template and filename
+     *
+     * @param string $template
+     * @param string $filename
+     * @return \Illuminate\Http\Response
+     */
+    protected function getImage($template, $filename)
+    {
+        // Get image path from cache or create new cached version
+        $path = ImageCache::getCachedImage(
+            $template, 
+            $filename, 
+            $this->maxSize, 
+            $this->coords, 
+            $this->ratio
+        );
 
-    switch (true) {
-    // closure template found
-    case is_callable($template):
-      return $template;
+        if (!$path) {
+            abort(404);
+        }
 
-    // filter template found
-    case class_exists($template):
-      return new $template($this->maxSize, $this->coords, $this->ratio);
-
-    default:
-      // template not found
-      abort(404);
-      break;
+        // Return image with proper headers
+        return Response::file($path, [
+            'Content-Type' => mime_content_type($path),
+            'Cache-Control' => 'public, max-age=31536000',
+        ]);
     }
-  }
+
+    /**
+     * Get the original image
+     *
+     * @param string $filename
+     * @return \Illuminate\Http\Response
+     */
+    protected function getOriginal($filename)
+    {
+        foreach (Config::get('imagecache.paths') as $path) {
+            $image_path = $path . '/' . $filename;
+            if (file_exists($image_path)) {
+                return Response::file($image_path, [
+                    'Content-Type' => mime_content_type($image_path),
+                    'Cache-Control' => 'public, max-age=31536000',
+                ]);
+            }
+        }
+        abort(404);
+    }
+
+    /**
+     * Get the image as download
+     *
+     * @param string $filename
+     * @return \Illuminate\Http\Response
+     */
+    protected function getDownload($filename)
+    {
+        foreach (Config::get('imagecache.paths') as $path) {
+            $image_path = $path . '/' . $filename;
+            if (file_exists($image_path)) {
+                return Response::download($image_path);
+            }
+        }
+        abort(404);
+    }
 }
