@@ -1,27 +1,40 @@
 <?php
 namespace App\Providers;
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Cache\RateLimiting\Limit;
+use Illuminate\Support\Facades\RateLimiter;
+use App\Services\ImageCache;
 
 class AppServiceProvider extends ServiceProvider
 {
-  /**
-   * Register any application services.
-   *
-   * @return void
-   */
-  public function register()
-  {
-    //
-  }
+    /**
+     * Register any application services.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        // Register ImageCache service
+        $this->app->singleton('image-cache', function ($app) {
+            return new ImageCache();
+        });
+    }
 
-  /**
-   * Bootstrap any application services.
-   *
-   * @return void
-   */
-  public function boot()
-  {
-    setLocale(LC_ALL, 'de');
-    \Carbon\Carbon::setLocale('de');
-  }
+    /**
+     * Bootstrap any application services.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        setLocale(LC_ALL, 'de');
+        \Carbon\Carbon::setLocale('de');
+
+        // Configure rate limiter for API
+        RateLimiter::for('api', function ($request) {
+            return $request->user()
+                ? Limit::perMinute(60)->by($request->user()->id)
+                : Limit::perMinute(30)->by($request->ip());
+        });
+    }
 }
